+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @plim/core
 
+## 0.1.3
+
+### Patch Changes
+
+- Patch bump to keep all @plim packages on the same version for a synchronized release.
+
 ## 0.1.2
 
 ### Patch Changes